refactor: extract conversation handlers to hppk
